# api_views/books.py
import jsonschema

from api_views.users import token_validator, error_message_helper
from config import db
from api_views.json_schemas import *
from flask import jsonify, Response, request, json
from models.user_model import User
from models.books_model import Book


def get_all_books():
    return_value = jsonify({'Books': Book.get_all_books()})
    return return_value


def add_new_book():
    request_data = request.get_json()
    try:
        jsonschema.validate(request_data, add_book_schema)
    except:
        return Response(error_message_helper("Please provide a proper JSON body."), 400, mimetype="application/json")
    resp = token_validator(request.headers.get('Authorization'))
    if "error" in resp:
        return Response(error_message_helper(resp), 401, mimetype="application/json")
    else:
        user = User.query.filter_by(username=resp['sub']).first()

        # check if user already has this book title
        book = Book.query.filter_by(user=user, book_title=request_data.get('book_title')).first()
        if book:
            return Response(error_message_helper("Book Already exists!"), 400, mimetype="application/json")
        else:
            newBook = Book(book_title=request_data.get('book_title'), secret_content=request_data.get('secret'),
                           user_id=user.id)
            db.session.add(newBook)
            db.session.commit()
            responseObject = {
                'status': 'success',
                'message': 'Book has been added.'
            }
            return Response(json.dumps(responseObject), 200, mimetype="application/json")


def get_by_title(book_title):
    resp = token_validator(request.headers.get('Authorization'))
    if "error" in resp:
        return Response(error_message_helper(resp), 401, mimetype="application/json")
    else:
        user = User.query.filter_by(username=resp['sub']).first()
        book = Book.query.filter_by(user=user, book_title=str(book_title)).first()
        if book:
            responseObject = {
                'book_title': book.book_title,
                'secret': book.secret_content,
                'owner': book.user.username
            }
            return Response(json.dumps(responseObject), 200, mimetype="application/json")
        else:
            return Response(error_message_helper("Book not found!"), 404, mimetype="application/json")


def search_books():
    keyword = request.args.get('q', '', type=str)
    if not keyword or len(keyword) > 100:
        return Response(error_message_helper("Please provide a search keyword (max 100 chars)"), 400, mimetype="application/json")
    results = Book.query.filter(Book.book_title.contains(keyword)).all()
    books_list = [Book.json(book) for book in results]
    result = {
        'keyword': keyword,
        'count': len(books_list),
        'books': books_list
    }
    return Response(json.dumps(result), 200, mimetype="application/json")